split pytest and unittest tests
